[Release][1.0.5] renewal app

+++ core/network/src/main/kotlin/com/gradation/lift/network/dto/history/HistoryDto.kt
@@ -15,11 +15,11 @@ data class HistoryDto(
     @SerialName("progress")
     val progress: Int,
     @SerialName("work_time")
-    val workTime: Float,
+    val workTime: Int,
     @SerialName("rest_time")
-    val restTime: Float,
+    val restTime: Int,
     @SerialName("total_time")
-    val totalTime: Float,
+    val totalTime: Int,
     @SerialName("history_time_stamp")
     val historyTimeStamp: String,
     @SerialName("history_routine")
